requirements.txt
Flask==2.3.3
gunicorn==21.2.0
pycryptodome==3.19.0